Refactor account view and related components.
Extracted `FollowedAnimesRow` and `FollowedEpisodesRow` into dedicated files for better modularity. Removed unused `AccountCard` component and replaced it with customized cards. Updated `CustomCard` structure by eliminating margin and added Dart SDK version bump to `>=3.8.0`. Cleaned up related imports.
